// src/enums/qwallet-webhook-event.enum.ts
export enum QwalletWebhookEvent {
  WALLET_UPDATED = 'wallet-updated',
  WALLET_ADDRESS_GENERATED = 'wallet-address-generated',
  DEPOSIT_CONFIRMATION = 'deposit-confirmation',
  DEPOSIT_SUCCESSFUL = 'deposit-successful',
  DEPOSIT_ON_HOLD = 'deposit-on-hold',
  DEPOSIT_FAILED_AML = 'deposit-failed-aml',
  DEPOSIT_REJECTED = 'deposit-rejected',
  WITHDRAW_SUCCESSFUL = 'withdraw-successful',
  WITHDRAW_REJECTED = 'withdraw-rejected',
  ORDER_DONE = 'order-done',
  ORDER_CANCELLED = 'order-cancelled',
  SWAP_COMPLETED = 'swap-completed',
  SWAP_REVERSED = 'swap-reversed',
  SWAP_FAILED = 'swap-failed',
  SELL_SUCCESSFUL = 'sell-transaction-successful',
  SELL_PROCESSING = 'sell-transaction-processing',
  SELL_FAILED = 'sell-transaction-failed',
  BUY_SUCCESSFUL = 'buy-transaction-successful',
  BUY_PROCESSING = 'buy-transaction-processing',
  BUY_FAILED = 'buy-transaction-failed',
}